.Legacy/Python_Script/Python_Script/requirements.txt
attrs==23.2.0
bcrypt==4.1.3
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.8
exceptiongroup==1.2.1
iniconfig==2.0.0
mysql-connector-python==8.4.0
packaging==24.0
paramiko==3.4.0
pip==24.0
pluggy==1.5.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.2.2
setuptools==70.0.0
tomli==2.0.1
vboxapi==1.0
